+++ apps/nextjs-app/src/backend/api/graphql-sdl/graphql-sdl-context.ts
@@ -1,4 +1,4 @@
-import type { PrismaClientDbMain } from '@your-org/db-main-prisma';
+import type { PrismaClientDbMain } from '@we-make-websites/db-main-prisma';
 import { prismaClient } from '@/backend/config/container.config';
 
 export type GraphqlSdlContext = {