+++ apps/nextjs-app/src/pages/api/hello.ts
@@ -1,4 +1,4 @@
-import { sayHello } from '@your-org/core-lib';
+import { sayHello } from '@we-make-websites/core-lib';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 export default async function handleApiHelloRoute(