+++ packages/db-main-prisma/e2e/suites/prisma-cli-commands.test.ts
@@ -1,6 +1,9 @@
 import execa from 'execa';
 import type { Options as ExecaOptions } from 'execa';
-import { PrismaManager, PrismaClientDbMain } from '@your-org/db-main-prisma';
+import {
+  PrismaManager,
+  PrismaClientDbMain,
+} from '@we-make-websites/db-main-prisma';
 import { getAndCheckDatabaseDsn } from '../e2e-dsn-services.util';
 
 describe('prisma cli commands', () => {